CSS-in-JS: Dropdown button
