+++ src/dropdownButton/DropdownButton.unit.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { shallow } from 'enzyme'
 
-import DropdownButton from 'dropdownButton'
+import DropdownButton from './DropdownButton'
 import ChevronIcon from 'icon/chevronIcon'
 
 const defaultProps = {
