+++ src/dropdownButton/index.tsx
@@ -1,40 +1,33 @@
-import React from 'react'
-import cc from 'classcat'
-import style from './style'
-import { ChevronIcon } from 'icon'
-import { color } from '_utils/branding'
+import styled from 'styled-components'
 
-export interface DropdownButtonProps {
-  onClick: (event: React.MouseEvent<HTMLElement>) => void
-  children: JSX.Element | string
-  open?: boolean
-  className?: string
-  iconPosition?: 'left' | 'right'
-}
+import DropdownButton from './DropdownButton'
+import { color, font, transition } from '_utils/branding'
 
-const DropdownButton = ({
-  open = false,
-  children,
-  onClick,
-  className = '',
-  iconPosition = 'right',
-}: DropdownButtonProps) => (
-  <div
-    className={cc([
-      'kirk-dropdownButton',
-      {
-        'kirk-dropdownButton--open': open,
-      },
-      className,
-    ])}
-  >
-    <button aria-expanded={open} type="button" onClick={onClick}>
-      {iconPosition === 'left' && <ChevronIcon iconColor={color.icon} down />}
-      {children}
-      {iconPosition === 'right' && <ChevronIcon iconColor={color.icon} down />}
-    </button>
-    <style jsx>{style}</style>
-  </div>
-)
+const StyledDropdownButton = styled(DropdownButton)`
+  & {
+    position: relative;
+    display: inline-block;
+  }
 
-export default DropdownButton
+  & > button {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    color: ${color.primary};
+    background: none;
+    border: 0;
+    outline: 0;
+    font-size: ${font.base.size};
+    cursor: pointer;
+  }
+
+  & .kirk-icon {
+    transition: transform ${transition.duration.base} ease-out;
+  }
+
+  &.kirk-dropdownButton--open .kirk-icon {
+    transform: rotate(180deg);
+  }
+`
+
+export default StyledDropdownButton
